refactor(nutrition): extract search orchestration
